# better concurrency on resolver cache